@@ -1,26 +1,26 @@
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-import 'Provider/notification.dart';
-import 'main-page/main_page.dart';
+import 'package:flutter_auth/Screens/forgot_page.dart';
+import 'package:flutter_auth/Screens/login_page.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'Screens/auth_page.dart';
+import 'firebase_options.dart';
 
-void main() {
+void main() async{
+  WidgetsFlutterBinding.ensureInitialized();
+  await Firebase.initializeApp(
+    options: DefaultFirebaseOptions.currentPlatform,
+  );
   runApp(const MyApp());
-}
+} 
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
+
   @override
   Widget build(BuildContext context) {
-    return MultiProvider(
-      providers: [
-        ChangeNotifierProvider(
-          create: (context) => NotiCounter(),
-        )
-      ],
-      child: const MaterialApp(
-        debugShowCheckedModeBanner: false,
-        home: MainPage(),
-      ),
+    return const MaterialApp(
+      debugShowCheckedModeBanner: false,
+      home: AuthPage(),
     );
   }
 }
