@@ -1,4 +1,4 @@
-package com.example.clone_linkin
+package com.example.flutter_auth
 
 import io.flutter.embedding.android.FlutterActivity
 
