@@ -3,7 +3,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:fluttericon/font_awesome5_icons.dart';
 import 'package:hexcolor/hexcolor.dart';
-
+import 'package:firebase_auth/firebase_auth.dart';
 import '../main_page.dart';
 
 class SettingPage extends StatefulWidget {
@@ -14,6 +14,7 @@ class SettingPage extends StatefulWidget {
 }
 
 class _SettingPageState extends State<SettingPage> {
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -23,7 +24,7 @@ class _SettingPageState extends State<SettingPage> {
           onPressed: () {
             Navigator.push(
               context,
-              MaterialPageRoute(builder: (context) => const MainPage()),
+              MaterialPageRoute(builder: (context) => MainPage()),
             );
           },
           icon: const Icon(
@@ -152,10 +153,11 @@ class _SettingPageState extends State<SettingPage> {
                     ),
                   ),
                   InkWell(
-                    onTap: () {},
+                    onTap: (){},
                     child: Container(
                       width: double.maxFinite,
                       height: 30,
+                      
                       alignment: Alignment.centerLeft,
                       child: const Text(
                         "Sign Out",
